Remove choice around SoundCloud URL parameters
Long term issue fix for issue arising from PR #3 and fixed in #4.

The specific choices for query parameters in sound cloud links vastly affect
the functionality of the page, indeed, sometimes breaking it. This removes
the choice of the user being able to set the full URL in the _data and instead
only prompts them to provide a track_id